Add JSDoc tags in the comments.
It describes the types of every parameter of a function, and also the
types of the properties of a class.
It improves coding efficiency in modern IDEs that are able to interpret
these tags.
It works pretty well with Netbeans 8.1 beta (except the callback
parameters)

+ small fix for the language property of an archive (mix between
"language" and "_language" in the code)