Fix in the service worker that was hanging in Firefox OS on many pages.
It seems that Firefox OS is disturbed when it has a reject or no answer
on some HTTP requests, where Firefox desktop and Chromium simply ignore
that.

This fix sends a dummy content on javascript requests, which allows
Firefox OS to display at least some articles of the Ray Charles file.
But some articles still can't be read, probably because of some other
missing dependencies.