Fix a bug in unit tests, that appeared with recent versions of browsers.
Recent versions of Firefox and Chromium did not like the way I was
trying to find the webworker_bzip2.js file in 2 different possible locations
(because the path is slightly different when running from a unit test)
I changed this hack by another one : use a global variable that stores the
prefix of this path, with a different value in a unit test.
There should be a better way to do that, but this is only needed for legacy
Evopedia archives. So it's not worth spending too much time on things that
are now discontinued.