Handle subdirectories in article namespace.
Works in jQuery mode, not yet in ServiceWorker mode.
Fixes #205
It involved to avoid making the (wrong) assumption that
there all articles are directly under the A namespace.
We use the "base" tag to specify the current path.
It allowed to remove some unclean code in zimArchive.js

Also changed the signature of the callback of readArticle,
so that to have the whole dirEntry if necessary in the future.