More complex scenario for UI tests.
It involves following links in the ZIM file, using the back button,
and checking for CSS styles and images.
Some versions of Edge have been added to the list of tested browsers.
A status of these tests is also added in the README.md

The bottom bar had to be hidden for these tests to work.
Some UI tests did not work on some versions of Edge, and have been disabled.