tests: Run "sauce" job for same-repo push instead of push+pr
There is no need to run the job on both since one always implies
the other. A same-repo PR by definition has every commit in it
trigger a push.

Triggering the job both ways is not only inefficient, but means
SauceLabs connection end up refused during the Nightwatch test
since it will find its free concurrency limit being reached from
the two parallel versions of the same job.

The Karma version seems to not suffer from this as it re-tries
and merely waits a little bit. We may encounter this issue more
in the future e.g. from unrelated commits being worked on, but
alas, this is an improvement either way.